[Calling][Feature] Rename NavigationBarViewData to SetupScreenViewData (#439)
