+++ azure-communication-ui/calling/src/main/java/com/azure/android/communication/ui/calling/presentation/fragment/setup/SetupFragment.kt
@@ -133,8 +133,8 @@ internal class SetupFragment :
         }
 
         val localOptions = holder.container.configuration.callCompositeLocalOptions
-        val titleSpan = if (!TextUtils.isEmpty(localOptions?.navigationBarViewData?.title)) {
-            SpannableString(localOptions?.navigationBarViewData?.title)
+        val titleSpan = if (!TextUtils.isEmpty(localOptions?.setupScreenViewData?.title)) {
+            SpannableString(localOptions?.setupScreenViewData?.title)
         } else {
             SpannableString(getString(R.string.azure_communication_ui_calling_call_setup_action_bar_title))
         }
@@ -144,10 +144,10 @@ internal class SetupFragment :
         callCompositeActivity.supportActionBar?.title = titleSpan
 
         // Only set the subtitle if the title has also been set
-        if (!TextUtils.isEmpty(localOptions?.navigationBarViewData?.title) &&
-            !TextUtils.isEmpty(localOptions?.navigationBarViewData?.subtitle)
+        if (!TextUtils.isEmpty(localOptions?.setupScreenViewData?.title) &&
+            !TextUtils.isEmpty(localOptions?.setupScreenViewData?.subtitle)
         ) {
-            val subtitleSpan = SpannableString(localOptions?.navigationBarViewData?.subtitle)
+            val subtitleSpan = SpannableString(localOptions?.setupScreenViewData?.subtitle)
             setActionbarTextColor(subtitleSpan, R.color.azure_communication_ui_calling_color_action_bar_subtext)
             callCompositeActivity.supportActionBar?.subtitle = subtitleSpan
         }
