+++ azure-communication-ui/demo-app/src/call-with-chat/java/com/azure/android/communication/ui/callwithchatdemoapp/launcher/CallWithChatCompositeJavaLauncher.java
