+++ azure-communication-ui/demo-app/src/calling/java/com/azure/android/communication/ui/callingcompositedemoapp/launcher/CallingCompositeJavaLauncher.java
@@ -11,8 +11,8 @@ import com.azure.android.communication.ui.calling.models.CallCompositeGroupCallL
 import com.azure.android.communication.ui.calling.models.CallCompositeJoinLocator;
 import com.azure.android.communication.ui.calling.models.CallCompositeLocalOptions;
 import com.azure.android.communication.ui.calling.models.CallCompositeLocalizationOptions;
-import com.azure.android.communication.ui.calling.models.CallCompositeNavigationBarViewData;
 import com.azure.android.communication.ui.calling.models.CallCompositeRemoteOptions;
+import com.azure.android.communication.ui.calling.models.CallCompositeSetupScreenViewData;
 import com.azure.android.communication.ui.calling.models.CallCompositeTeamsMeetingLinkLocator;
 import com.azure.android.communication.ui.callingcompositedemoapp.CallLauncherActivity;
 import com.azure.android.communication.ui.callingcompositedemoapp.CallLauncherActivityErrorHandler;
@@ -79,9 +79,9 @@ public class CallingCompositeJavaLauncher implements CallingCompositeLauncher {
         final CallCompositeLocalOptions localOptions = new CallCompositeLocalOptions()
                 .setParticipantViewData(SettingsFeatures
                         .getParticipantViewData(callLauncherActivity.getApplicationContext()))
-                .setNavigationBarViewData(
-                        new CallCompositeNavigationBarViewData(SettingsFeatures.getTitle())
-                                .setSubtitle(SettingsFeatures.getSubtitle()));
+                .setSetupScreenViewData(
+                        new CallCompositeSetupScreenViewData()
+                            .setTitleAndSubtitle(SettingsFeatures.getTitle(), SettingsFeatures.getSubtitle()));
 
         callComposite.launch(callLauncherActivity, remoteOptions, localOptions);
     }
