+++ docs/CHANGELOG_UI_CALLING.md
@@ -3,16 +3,16 @@
 ## 1.1.0-beta.1 (2022-09-20)
 
 ### Features
-- Setting up Group Call Title and Subtitle is now possible by customizing CallCompositeLocalOptions with CallCompositeNavigationBarViewData
+- Setting up Call Title and Subtitle is now availble by customizing `CallCompositeLocalOptions` with `CallCompositeSetupScreenViewData`.
 - Implemented new error message `cameraFailure` that can be sent to developers when initiating or turning on camera fails.
 - Error message now shown when network is not available before joining a call.
 - Added new button to allow user to quickly navigate to app's info page when permissions are denied.
 
 ### Bug Fixes
-- Display DrawerDialog across screen rotation
-- Fix ANR when trying to hang up call on hold
-- Fix edge case with multiple activity instances
-- Fix display name not getting truncated in participant list when they are too long (https://github.com/Azure/communication-ui-library-android/pull/370)
+- Display DrawerDialog across screen rotation.
+- Fix ANR when trying to hang up call on hold.
+- Fix edge case with multiple activity instances.
+- Fix display name not getting truncated in participant list when they are too long (https://github.com/Azure/communication-ui-library-android/pull/370).
 
 ## 1.0.0 (2022-06-20)
 - This version is the public GA release with Calling UI Library
